requirements.txt
accelerate==0.21.0
aiodogstatsd==0.16.0.post0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
black==23.3.0
bzt==1.16.22
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorlog==6.7.0
cryptography==41.0.1
cssselect==1.2.0
Cython==0.29.34
dill==0.3.6
diskcache==5.6.1
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
fuzzyset2==0.2.2
gptcache==0.1.30
guidance==0.0.63
hdrpy==0.3.3
huggingface-hub==0.16.4
humanize==4.6.0
idna==3.4
influxdb==5.3.1
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
molotov==2.6
mpmath==1.3.0
msal==1.22.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
openai==0.27.9
packaging==23.1
parsimonious==0.10.0
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.5.1
progressbar33==2.4
prompt-toolkit==3.0.38
psutil==5.9.5
pycparser==2.21
pygtrie==2.5.0
PyJWT==2.7.0
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
pytz==2023.3
PyVirtualDisplay==3.0
PyYAML==6.0
rapidfuzz==3.0.0
regex==2023.5.5
requests==2.30.0
ruff==0.0.285
safetensors==0.3.1
six==1.16.0
sympy==1.12
terminaltables==3.1.10
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.5.0
urllib3==2.0.2
urwid==2.1.2
wcwidth==0.2.6
websocket-client==1.5.2
xformers==0.0.20
yarl==1.9.2
